refactor(model-server): drop the HTML OpenAPI spec
There's not much sense in maintaining a "REST" API specification for a
server-side-rendered HTML frontend.
