+++ model-server/src/main/kotlin/org/modelix/model/server/handlers/HistoryHandler.kt
@@ -4,9 +4,9 @@ import io.ktor.server.application.Application
 import io.ktor.server.application.call
 import io.ktor.server.html.respondHtmlTemplate
 import io.ktor.server.request.receiveParameters
-import io.ktor.server.resources.get
-import io.ktor.server.resources.post
 import io.ktor.server.response.respondRedirect
+import io.ktor.server.routing.get
+import io.ktor.server.routing.post
 import io.ktor.server.routing.routing
 import kotlinx.datetime.Instant
 import kotlinx.datetime.TimeZone
@@ -34,7 +34,6 @@ import kotlinx.html.thead
 import kotlinx.html.tr
 import kotlinx.html.ul
 import kotlinx.html.unsafe
-import org.modelix.api.html.Paths
 import org.modelix.authorization.KeycloakScope
 import org.modelix.authorization.asResource
 import org.modelix.authorization.getUserName
@@ -60,10 +59,10 @@ class HistoryHandler(val client: IModelClient, private val repositoriesManager: 
 
     fun init(application: Application) {
         application.routing {
-            get<Paths.getHistory> {
+            get("/history") {
                 call.respondRedirect("../repos/")
             }
-            get<Paths.getRepoAndBranch> {
+            get("/history/{repoId}/{branch}") {
                 val repositoryId = RepositoryId(call.parameters["repoId"]!!)
                 val branch = repositoryId.getBranchReference(call.parameters["branch"]!!)
                 val params = call.request.queryParameters
@@ -92,7 +91,7 @@ class HistoryHandler(val client: IModelClient, private val repositoriesManager: 
                 }
             }
             requiresPermission("history".asResource(), KeycloakScope.WRITE) {
-                post<Paths.revertBranch> {
+                post("/history/{repoId}/{branch}/revert") {
                     val repositoryId = RepositoryId(call.parameters["repoId"]!!)
                     val branch = repositoryId.getBranchReference(call.parameters["branch"]!!)
                     val params = call.receiveParameters()
